server/requirements.txt
bcrypt>=4.0.0
Flask>=2.0.0
Flask-Cors>=3.0.0
Flask-JWT-Extended>=4.0.0
mysql-connector-python>=8.0.0
cryptography>=3.0.0
python-dotenv>=0.19.0
Flask-Mail>=0.9.1
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
reportlab>=3.6.0